+++ src/main/java/io/github/sanyarnd/applocker/MessageHandler.java
@@ -15,7 +15,7 @@
 //     See the License for the specific language governing permissions and
 //     limitations under the License.
 
-package io.github.sanyarnd.applocker.messaging;
+package io.github.sanyarnd.applocker;
 
 import java.io.Serializable;