+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-# 1.0.1
+# 1.0.2
+- `AppLocker.Builder`'s `busy` method now supports `Runnable` callback
+- javadoc and sources are now distributed along with the package 
 
+# 1.0.1
 - Removed unnecessary builders
 - Improved test coverage
 
 # 1.0
-
 - Initial commit